// api/src/exceptions/exception-code.enum.ts
export const enum ExceptionCodeEnum {
    MYSQL_SERVICE__QUERY_ERR = 'MYSQ-SERVICE-000-QRERR',
    MYSQL_SERVICE__CONN_ERR = 'MYSQL-SERVICE-001-CNERR',
    VALIDATION_UTIL__ERR = 'VALID-001-ERR',
    EXPRESS_APP__START_ERR = 'EXPRESS-001-ERR',
    CRON_TASK__GENERAL_PROCEEDING_ERR = 'CRON-TASK-PROCEEDING-003-TSERR',
    AXIOS__ERR = 'AXIOS-REQ__001-ERR',
    BASE_SERVICE__GENERAL_ERR = 'BASE-SERVICE-000-GENERR',
    PROXY_SERVICE__UNEXPECTED_STATUS = 'PROXY-SERVICE-000-STERR',
    INTEGRATION_SERVICE__GENERAL_ERR = 'INTEGRATION-SERVICE-000-GENERR',
    BASELINKER_SERVICE__GENERAL_ERR = 'BASELINKER-SERVICE-000-GENERR',
    APILO_SERVICE__GENERAL_ERR = 'APILO-SERVICE-000-GENERR',
    WOO_SERVICE__GENERAL_ERR = 'WOO-SERVICE-000-GENERR',
    PRESTA_SERVICE__GENERAL_ERR = 'PRESTA-SERVICE-000-GENERR',
    ADAPTER_FACTORY__GENERAL_ERR = 'ADAPTER_FACTORY-000-GENERR',
    PROPAGATOR__GENERAL_ERR = 'PROPAGATOR-000-GENERR',
    EBIUROPROXY__GENERAL_ERR = 'EBIUROPROXY-000-GENERR',
}