// api/src/api-server/baselinker/enums.ts
export enum BaselinkerMethods {
    GET_INVENTORIES = 'getInventories',
    GET_INVENTORY_PRODUCTS_LIST = 'getInventoryProductsList',
    GET_INVENTORY_PRODUCTS_DATA = 'getInventoryProductsData',
    GET_STORAGES_LIST = 'getStoragesList',
    GET_PRODUCTS_DATA = 'getProductsData',
    GET_ORDERS = 'getOrders',
    GET_CATEGORIES = 'getCategories',
    ADD_CATEGORY = 'addCategory',
    ADD_PRODUCT = 'addProduct',
    DELETE_PRODUCT = 'deleteProduct',
}
export enum BaseLinkerStatusResponse {
    SUCCESS = 'SUCCESS',
    ERROR = 'ERROR'
}

export enum BaseLinkerStorageId {
    BL = 'bl_1'
}